Allow -f to take a value from the following argument
This allows the -f option to take its value from the next command line argument, instead of requiring -f and the address to be run together into a single argument.  Real sendmail allows both syntaxes, and when MediaWiki invokes PHP's mail() function, it does it with the separate-argument syntax (so that unpatched mini_sendmail cannot be used for MediaWiki).  Unfortunately, even with this patch it still doesn't work in my MediaWiki installation, but adding the missing syntax at least removes one of the obstacles.